move roundDown and numberFormat utils.js and import at origin
